db+sqlite+pg: convert NewImportedAddress to db-ops interface style
Introduce a NewImportedAddressOps interface (WalletWatchOnly,
CreateImportedAddress, InsertAddressSecret) and a NewImportedAddressWithOps
workflow in new addressstore_newimported.go files.
