+++ wallet/internal/db/addresses_common.go
@@ -43,74 +43,6 @@ var (
 	)
 )
 
-// ValidateBasic checks imported address creation parameters that do not depend
-// on external state.
-func (p NewImportedAddressParams) ValidateBasic() error {
-	if len(p.ScriptPubKey) == 0 {
-		return ErrMissingScriptPubKey
-	}
-
-	return nil
-}
-
-// ValidateWatchOnly checks imported address creation parameters against the
-// parent wallet's watch-only state. A watch-only wallet must not receive
-// private-key-bearing imports. The symmetric direction (a spendable wallet
-// must not receive an imported address without private-key material) is
-// enforced at the SQL-backend entry through requireAddressPrivKeyOnSpendable;
-// kvdb's data model carries different invariants around legacy imported
-// address rows (a grandfathered legacy shape), so the symmetric check is not
-// applied there.
-func (p NewImportedAddressParams) ValidateWatchOnly(
-	walletIsWatchOnly bool) error {
-
-	if walletIsWatchOnly && p.HasPrivateKey() {
-		return fmt.Errorf("watch-only wallet %d cannot import private-key-"+
-			"bearing address into account %q: %w",
-			p.WalletID, DefaultImportedAccountName, ErrWatchOnlyViolation)
-	}
-
-	return nil
-}
-
-// RequireAddressPrivKeyOnSpendable enforces the ADR 0012 symmetric
-// invariant for SQL backends: a spendable wallet must not receive an
-// imported address without encrypted private-key material. Public-only
-// and script-only imports are both rejected (HasPrivateKey covers both
-// — script-only imports have material in EncryptedScript but no priv
-// key).
-func RequireAddressPrivKeyOnSpendable(walletID uint32,
-	walletIsWatchOnly bool, hasPrivKey bool) error {
-
-	if walletIsWatchOnly || hasPrivKey {
-		return nil
-	}
-
-	return fmt.Errorf("spendable wallet %d cannot import address "+
-		"without private-key material into account %q: %w",
-		walletID, DefaultImportedAccountName,
-		ErrSpendableWalletNeedsAddressPrivKey)
-}
-
-// HasPrivateKey returns true if the params include private key material.
-// Script material is tracked separately and does not make an imported address
-// spend-capable by itself.
-func (p NewImportedAddressParams) HasPrivateKey() bool {
-	return len(p.EncryptedPrivateKey) > 0
-}
-
-// HasSecretMaterial returns true when the imported address carries any
-// encrypted secret payload that should be persisted in address_secrets.
-// Private keys and scripts are stored independently.
-func (p NewImportedAddressParams) HasSecretMaterial() bool {
-	return len(p.EncryptedPrivateKey) > 0 || len(p.EncryptedScript) > 0
-}
-
-// HasScript returns true if the params include script spend data.
-func (p NewImportedAddressParams) HasScript() bool {
-	return len(p.EncryptedScript) > 0
-}
-
 // DerivedAddressAdapters groups the functions needed to create a
 // derived address across different database backends.
 type DerivedAddressAdapters[QTX any, AccountRow any, AccountParams any,
